+++ tests/snapshot_tests/test_snapshots.py
@@ -1275,6 +1275,11 @@ def test_data_table_in_tabs(snap_compare):
     assert snap_compare(SNAPSHOT_APPS_DIR / "data_table_tabs.py")
 
 
+def test_auto_tab_active(snap_compare):
+    # https://github.com/Textualize/textual/issues/4593
+    assert snap_compare(SNAPSHOT_APPS_DIR / "auto_tab_active.py", press=["space"])
+
+
 def test_footer_space_between_key_and_description(snap_compare):
     """Regression test for https://github.com/Textualize/textual/issues/4557"""
     assert snap_compare(SNAPSHOT_APPS_DIR / "footer_spacing.py")