+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.65.0] - 2024-06-05
 
 ### Fixed
 
@@ -2061,6 +2061,7 @@ https://textual.textualize.io/blog/2022/11/08/version-040/#version-040
 - New handler system for messages that doesn't require inheritance
 - Improved traceback handling
 
+[0.65.0]: https://github.com/Textualize/textual/compare/v0.64.0...v0.65.0
 [0.64.0]: https://github.com/Textualize/textual/compare/v0.63.6...v0.64.0
 [0.63.6]: https://github.com/Textualize/textual/compare/v0.63.5...v0.63.6
 [0.63.5]: https://github.com/Textualize/textual/compare/v0.63.4...v0.63.5